Fixed #1979 - Refactor Retention Index Library Service

+++ chemclipse/plugins/org.eclipse.chemclipse.model/src/org/eclipse/chemclipse/model/services/IRetentionIndexLibraryService.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2023 Lablicate GmbH.
+ * Copyright (c) 2023, 2024 Lablicate GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -24,5 +24,5 @@ public interface IRetentionIndexLibraryService {
 
 	String getVersion();
 
-	List<ILibraryInformation> getLibraryInformation(float retentionIndex, RetentionIndexLibrarySettings settings, IProgressMonitor monitor);
+	List<ILibraryInformation> getLibraryInformation(int retentionIndex, RetentionIndexLibrarySettings settings, IProgressMonitor monitor);
 }
